+++ ComputedTomographyRotationCT/ComputedTomographyRotationCT.cs
@@ -96,24 +96,24 @@ namespace FLImagingExamplesCSharp
 				if((res = computedTomographyRotationCT.SetObjectTotalRotationAngle(360.00)).IsFail())
 					break;
 
-				TPoint3<double> tpObjectLocation = new TPoint3<double>();
-				tpObjectLocation.x = 0.00;
-				tpObjectLocation.y = 0.00;
-				tpObjectLocation.z = 28.6333;
+				TPoint3<float> tpObjectLocation = new TPoint3<float>();
+				tpObjectLocation.x = 0.00f;
+				tpObjectLocation.y = 0.00f;
+				tpObjectLocation.z = 28.6333f;
 				if((res = computedTomographyRotationCT.SetObjectLocation(tpObjectLocation)).IsFail())
 					break;
 
-				TPoint3<double> tpObjectEulerAngle = new TPoint3<double>();
-				tpObjectEulerAngle.x = 35.00;
-				tpObjectEulerAngle.y = 0.00;
-				tpObjectEulerAngle.z = 0.00;
+				TPoint3<float> tpObjectEulerAngle = new TPoint3<float>();
+				tpObjectEulerAngle.x = 35.00f;
+				tpObjectEulerAngle.y = 0.00f;
+				tpObjectEulerAngle.z = 0.00f;
 				if((res = computedTomographyRotationCT.SetObjectEulerAngle(tpObjectEulerAngle)).IsFail())
 					break;
 
-				TPoint3<double> tpObjectVoxelSize = new TPoint3<double>();
-				tpObjectVoxelSize.x = 0.01;
-				tpObjectVoxelSize.y = 0.01;
-				tpObjectVoxelSize.z = 0.01;
+				TPoint3<float> tpObjectVoxelSize = new TPoint3<float>();
+				tpObjectVoxelSize.x = 0.01f;
+				tpObjectVoxelSize.y = 0.01f;
+				tpObjectVoxelSize.z = 0.01f;
 				if((res = computedTomographyRotationCT.SetObjectVoxelSize(tpObjectVoxelSize)).IsFail())
 					break;
